Health and damage are now doubles, not ints.

+++ src/main/java/com/gmail/nossr50/events/fake/FakeEntityDamageEvent.java
@@ -7,7 +7,7 @@ import org.bukkit.event.entity.EntityDamageEvent;
  * Called when mcMMO applies damage due to special abilities.
  */
 public class FakeEntityDamageEvent extends EntityDamageEvent {
-    public FakeEntityDamageEvent(Entity damagee, DamageCause cause, int damage) {
+    public FakeEntityDamageEvent(Entity damagee, DamageCause cause, double damage) {
         super(damagee, cause, damage);
     }
 }
